refactor: 내가 시청한 동영상, 내가 업로드한 동영상 리스트 컴포넌트 분리, hr태그 margin값 조정
